pyproject.toml
[project]
name = "chayabusaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "jinja2>=3.1.6",
    "python-evtx>=0.8.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "xmltodict>=0.14.2",
]
